Rename MailSender to MailSenderOld and MailSenderNew to MailSender
