--- core/src/it/java/nl/nn/adapterframework/mailsenders/MailSenderNewWrapper.java
@@ -1,7 +1,0 @@
-package nl.nn.adapterframework.mailsenders;
-
-import nl.nn.adapterframework.senders.MailSenderNew;
-
-public class MailSenderNewWrapper extends MailSenderNew implements IMailSender {
-
-}
