+++ core/src/it/java/nl/nn/adapterframework/mailsenders/MailSenderOldWrapper.java
@@ -1,0 +1,7 @@
+package nl.nn.adapterframework.mailsenders;
+
+import nl.nn.adapterframework.senders.MailSenderOld;
+
+public class MailSenderOldWrapper extends MailSenderOld implements IMailSender {
+
+}
