+++ core/src/it/java/nl/nn/adapterframework/mailsenders/MailSenderTest.java
@@ -1,14 +1,5 @@
 package nl.nn.adapterframework.mailsenders;
 
-import java.io.IOException;
-
-import org.junit.Ignore;
-import org.junit.Test;
-
-import nl.nn.adapterframework.configuration.ConfigurationException;
-import nl.nn.adapterframework.core.SenderException;
-import nl.nn.adapterframework.core.TimeOutException;
-
 public class MailSenderTest extends MailSenderTestBase<MailSenderWrapper> {
 	private String smtpHost = "localhost"; // could be smtp.sendgrid.net
 	private String userId = "";
@@ -22,13 +13,4 @@ public class MailSenderTest extends MailSenderTestBase<MailSenderWrapper> {
 		ms.setSmtpPassword(password);
 		return ms;
 	}
-
-	@Ignore("MailSender throws NullPointerException because of a bug")
-	@Test
-	@Override
-	public void testParametersEmptyRecipients()
-			throws SenderException, TimeOutException, ConfigurationException, IOException {
-		super.testParametersEmptyRecipients();
-	}
-
 }
